# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Inventory management system API for QwikShelf",
        "title": "QwikShelf API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "QwikShelf Support",
            "email": "support@qwikshelf.local"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/inventory": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns inventory levels, optionally filtered by warehouse",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inventory"
                ],
                "summary": "List inventory",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by warehouse ID",
                        "name": "warehouse_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_qwikshelf_api_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_qwikshelf_api_internal_application_dto.InventoryLevelResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/inventory/adjust": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Adjusts stock level for a product variant in a warehouse",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inventory"
                ],
                "summary": "Adjust inventory",
                "parameters": [
                    {
                        "description": "Adjustment details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_qwikshelf_api_internal_application_dto.AdjustInventoryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_qwikshelf_api_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_qwikshelf_api_internal_application_dto.InventoryLevelResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_qwikshelf_api_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_qwikshelf_api_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/inventory/transfer": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates an inventory transfer between warehouses",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inventory"
                ],
                "summary": "Transfer inventory",
                "parameters": [
                    {
                        "description": "Transfer details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_qwikshelf_api_internal_application_dto.CreateTransferRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_qwikshelf_api_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_qwikshelf_api_internal_application_dto.TransferResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_qwikshelf_api_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_qwikshelf_api_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/inventory/warehouse/{warehouseId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns inventory levels for a specific warehouse",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Inventory"
                ],
                "summary": "List inventory by warehouse",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Warehouse ID",
                        "name": "warehouseId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_qwikshelf_api_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_qwikshelf_api_internal_application_dto.InventoryLevelResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_qwikshelf_api_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_qwikshelf_api_internal_application_dto.AdjustInventoryRequest": {
            "type": "object",
            "required": [
                "quantity_delta",
                "variant_id",
                "warehouse_id"
            ],
            "properties": {
                "quantity_delta": {
                    "type": "number"
                },
                "reason": {
                    "type": "string"
                },
                "variant_id": {
                    "type": "integer"
                },
                "warehouse_id": {
                    "type": "integer"
                }
            }
        },
        "github_com_qwikshelf_api_internal_application_dto.CategoryResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_qwikshelf_api_internal_application_dto.CreateTransferRequest": {
            "type": "object",
            "required": [
                "destination_warehouse_id",
                "items",
                "source_warehouse_id"
            ],
            "properties": {
                "destination_warehouse_id": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/github_com_qwikshelf_api_internal_application_dto.TransferItemRequest"
                    }
                },
                "source_warehouse_id": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_qwikshelf_api_internal_application_dto.InventoryLevelResponse": {
            "type": "object",
            "properties": {
                "batch_number": {
                    "type": "integer"
                },
                "expiry_date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "number"
                },
                "variant": {
                    "$ref": "#/definitions/github_com_qwikshelf_api_internal_application_dto.ProductVariantResponse"
                },
                "variant_id": {
                    "type": "integer"
                },
                "warehouse": {
                    "$ref": "#/definitions/github_com_qwikshelf_api_internal_application_dto.WarehouseResponse"
                },
                "warehouse_id": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_qwikshelf_api_internal_application_dto.ProductFamilyResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/github_com_qwikshelf_api_internal_application_dto.CategoryResponse"
                },
                "category_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_qwikshelf_api_internal_application_dto.ProductVariantResponse": {
            "type": "object",
            "properties": {
                "barcode": {
                    "type": "string"
                },
                "cost_price": {
                    "type": "number"
                },
                "family": {
                    "$ref": "#/definitions/github_com_qwikshelf_api_internal_application_dto.ProductFamilyResponse"
                },
                "family_id": {
                    "type": "integer"
                },
                "family_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_manufactured": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "selling_price": {
                    "type": "number"
                },
                "sku": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
...
        "github_com_qwikshelf_api_internal_application_dto.TransferItemRequest": {
            "type": "object",
            "required": [
                "quantity",
                "variant_id"
            ],
            "properties": {
                "quantity": {
                    "type": "number"
                },
                "variant_id": {
                    "type": "integer"
                }
            }
        },
        "github_com_qwikshelf_api_internal_application_dto.TransferItemResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "number"
                },
                "variant": {
                    "$ref": "#/definitions/github_com_qwikshelf_api_internal_application_dto.ProductVariantResponse"
                },
                "variant_id": {
                    "type": "integer"
                }
            }
        },
        "github_com_qwikshelf_api_internal_application_dto.TransferResponse": {
            "type": "object",
            "properties": {
                "authorized_by_user_id": {
                    "type": "integer"
                },
                "destination_warehouse": {
                    "$ref": "#/definitions/github_com_qwikshelf_api_internal_application_dto.WarehouseResponse"
                },
                "destination_warehouse_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_qwikshelf_api_internal_application_dto.TransferItemResponse"
                    }
                },
                "source_warehouse": {
                    "$ref": "#/definitions/github_com_qwikshelf_api_internal_application_dto.WarehouseResponse"
                },
                "source_warehouse_id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/github_com_qwikshelf_api_internal_domain_entity.TransferStatus"
                },
                "transferred_at": {
                    "type": "string"
                }
            }
        },
...
        "github_com_qwikshelf_api_internal_application_dto.WarehouseResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/github_com_qwikshelf_api_internal_domain_entity.WarehouseType"
                }
            }
        },
        "github_com_qwikshelf_api_internal_domain_entity.TransferStatus": {
            "type": "string",
            "enum": [
                "PENDING",
                "IN_TRANSIT",
                "COMPLETED",
                "CANCELLED"
            ],
            "x-enum-varnames": [
                "TransferStatusPending",
                "TransferStatusInTransit",
                "TransferStatusCompleted",
                "TransferStatusCancelled"
            ]
        },
        "github_com_qwikshelf_api_internal_domain_entity.WarehouseType": {
            "type": "string",
            "enum": [
                "store",
                "factory",
                "distribution_center"
            ],
            "x-enum-varnames": [
                "WarehouseTypeStore",
                "WarehouseTypeFactory",
                "WarehouseTypeDistributionCenter"
            ]
        },
        "github_com_qwikshelf_api_pkg_response.ErrorInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_qwikshelf_api_pkg_response.Meta": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "github_com_qwikshelf_api_pkg_response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/github_com_qwikshelf_api_pkg_response.ErrorInfo"
                },
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/github_com_qwikshelf_api_pkg_response.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your JWT token with the `Bearer ` prefix, e.g. `Bearer eyJhbGci...`",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}